refactor: reestructuring project folders
