+++ src/app/components/organisms/Header.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import "./header.css";
+
+const Header = () => {
+  return (
+    <div className="static">
+      <div className="header">Header</div>
+    </div>
+  );
+};
+
+export default Header;
