+++ src/app/components/organisms/header.css
@@ -1,0 +1,7 @@
+@tailwind components;
+
+@layer components {
+  .header {
+    @apply absolute top-[50px];
+  }
+}
